backend/requirements.txt
fastapi
uvicorn
streamlit
psycopg2-binary